pyproject.toml
[tool.poetry]
name = "vectordb"
version = "0.1.0"
description = "A distributed vector db build with faiss"
authors = ["sviridov <sviridov@umich.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
click = "8.1.7"
faiss-cpu = "1.9.0"
fastapi = "0.115.3"
googleapis-common-protos = "1.65.0"
grpcio = "1.67.1"
grpcio-tools = "1.67.1"
h11 = "0.14.0"
idna = "3.10"
json-tricks = "3.17.3"
numpy = "2.1.2"
packaging = "24.1"
protobuf = "5.28.3"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
redis = "5.2.0"
setuptools = "75.3.0"
sniffio = "1.3.1"
starlette = "0.41.0"
typing-extensions = "4.12.2"
uvicorn = "0.32.0"
pyyaml = "6.0.2"

[tool.poetry.scripts]
coordinator = "vectordb.coordinator.__main__:main"
shard = "vectordb.shard.__main__:main"
maksdb = "vectordb.cli.cli:cli"
server = "vectordb.api.api:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
